requirements.txt
fastapi[all]
cachetools==5.3.1
fastapi-cache
fastapi-cache2
langkit==0.0.32
numpy==1.24.3
opencensus==0.11.4
opencensus_context==0.1.3
opencensus_ext_azure==1.1.13
presidio_analyzer==2.2.352
presidio_anonymizer==2.2.354
pycryptodome==3.20.0
pydantic==1.10.11
PyJWT==2.8.0
PyJWT==2.8.0
pyotp==2.9.0
pytz==2024.1
scikit_learn
SQLAlchemy==2.0.16
tqdm
Werkzeug==3.0.3
whylogs==1.3.32
azure.identity
azure-keyvault-secrets
gunicorn
Werkzeug
pyjwt[crypto]
psycopg2-binary
spacy
transformers
torch
torchvision
torchaudio
sentence_transformers
pyarrow
nltk
pyotp 
werkzeug
bcrypt
sqlalchemy